Consolidate components for card list
