--- ClientApp/components/CardsSegment.tsx
@@ -1,27 +1,0 @@
-import * as React from "react";
-import { Segment } from "semantic-ui-react";
-import { ToDoCard } from "./ToDoCard";
-import "../css/CardsSegment.css";
-
-interface ICardSegmentProps {
-    numCards: number;
-}
-
-export class CardSegment extends React.Component<ICardSegmentProps, {}> {
-
-    public createCards = () => {
-        let cards: Object[] = [];
-        for (let i = 0; i < this.props.numCards; i++) {
-            cards.push(<ToDoCard key={i}/>);
-        }
-        return cards;
-    }
-
-    public render() {
-        return (
-            <Segment className="cards-segment">
-                {this.createCards()}
-            </Segment>
-        );
-    }
-}
